+++ apps/_shared/ts/models/business-organisations.ts
@@ -1,0 +1,9 @@
+export interface BusinessOrganisationRowCSV {
+  organisationNumber: string
+  said: string
+  sboid: string
+  transportCompanyAbbreviation: string
+  transportCompanyBusinessRegisterName: string
+  descriptionDe: string
+  abbreviationDe: string
+}