+++ tools/atlas-geocode-lines/src/constants.ts
@@ -1,0 +1,77 @@
+import path from 'path';
+import * as OJP from 'ojp-sdk'; 
+
+const DATA_PATH = path.resolve('./data');
+
+export const ATLAS_LINE_CSV_PATH = (DATA_PATH + '/actual_date_line_versions_LATEST.csv');
+export const OJP_LIR_CACHE_PATH = (DATA_PATH + '/ojp_lir_cache.json');
+export const ATLAS_STOPS_GEOJSON_PATH = (DATA_PATH + '/atlas_stops.geojson');
+
+export const OJP_STAGE_CONFIG: OJP.StageConfig = {
+  key: 'someKey',
+  apiEndpoint: OJP.DEFAULT_STAGE.apiEndpoint,
+  authBearerKey: OJP.DEFAULT_STAGE.authBearerKey, // override with another key
+};
+
+// sleep interval between 2 OJP requests, the default key is limited to 50requests / minute
+// @see https://opentransportdata.swiss/en/limits-and-costs/
+export const OJP_REQUESTS_SLEEP_MS = 1200;
+
+export let DEBUG_slnid: string | null = null;
+// DEBUG_slnid = 'ch:1:slnid:1025759';
+
+export let DEBUG_Output_Names = false
+// DEBUG_Output_Names = true
+
+export let DEBUG_Row: string | null = 'Zone Seymaz-Voirons (Choulex, Gy, Jussy, Meinier, Presinge, Puplinge Collonge-Bellerive, Thônex, Vandoeuvres. Annemasse, Juvigny, Machilly, Saint-Cergues)'
+DEBUG_Row = null
+
+export const STOP_NAME_SEPARATOR = ' |SEP| ';
+
+interface SeparatorData {
+  regexp: RegExp,
+  description: string,
+  example: string,
+  replace: string
+}
+
+export const PSEUDO_SEPARATORS_DATA: SeparatorData[] = [
+  // order matters, keep general pseudo-separators on top
+  {
+    regexp: RegExp(/([a-zäöü]{2}), ([A-Z][a-zäöü]{1})/g),
+    description: 'comma ,',
+    example: 'Belmont, Bussigny-près-Lausanne, Chavannes-près-Renens, Crissier, Ecublens',
+    replace: '$1' + STOP_NAME_SEPARATOR + '$2',
+  },
+  {
+    regexp: RegExp(/([a-zäöü]{3})-([A-Z])/g),
+    description: 'dash -',
+    example: 'Romanshorn-Immenstaad-Hagnau-Altnau-Güttingen',
+    replace: '$1' + STOP_NAME_SEPARATOR + '$2',
+  },
+  {
+    regexp: RegExp(/([a-zäöü]{3})\/([A-Z])/g),
+    description: 'slash /',
+    example: 'Kriens Busschleife - Pilatus-Bahnen/Sidhalde/Sonnenberg',
+    replace: '$1' + STOP_NAME_SEPARATOR + '$2',
+  },
+];
+
+// Isolate stop names that can be split by the pseudoseparators above
+export const SEPARATOR_WORDS_TO_IGNORE = [
+  'Le Locle-Col-des-Roches',
+  'St-Gervais-les-Bains',
+  'Yverdon-les-Bains',
+  'Lancy-Pont-Rouge',
+  'Vers-Chez-les-Blanc',
+  'Zone du Haut-Plateau',
+  'Escher-Wyss-Platz',
+  'Lancy-Port-Rouge',
+  'Lancy-Pont-R.',
+  'Biel/Bienne',
+];
+
+export const STOP_NAMES_LOOKUP = [
+  'Schinznach Dorf', // Schinznach Dorf-Thalheim -> adds space -
+];
+