+++ tools/gtfs-static-db-importer/inc/sql/select_trips_group_by_stop_times.sql
@@ -1,6 +1,6 @@
 SELECT 
     trips.*,
-    GROUP_CONCAT(PRINTF('%s|%s|%s|%s', stop_times.ROWID, stop_times.stop_id, stop_times.arrival_time, stop_times.departure_time)) AS stop_times_data
+    GROUP_CONCAT(PRINTF('%s|%s|%s|%s', stop_times.ROWID, stop_times.stop_id, stop_times.arrival_time, stop_times.departure_time), ' -- ') AS stop_times_data
 FROM trips, stop_times
 WHERE
     trips.trip_id = stop_times.trip_id