+++ tools/scripts/cli_otd_fetch_atlas.sh
@@ -1,0 +1,16 @@
+DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
+LOGS_BASEPATH=$DIR/logs
+
+# DATE_NOW=$(date +"%Y-%m-%d-%H%M")
+DATE_NOW=$(date +"%Y-%m-%d")
+
+source $DIR/common.sh
+source $PYTHON_VENV_PATH/bin/activate
+
+ATLAS_FETCH_LOGFILE=$LOGS_BASEPATH/otd_fetch_csv_atlas-lines-$DATE_NOW.log
+python3 $DIR/cli_otd_fetch_csv_package.py --package_id slnid-line 2>&1 | tee $ATLAS_FETCH_LOGFILE
+symlink_latest $ATLAS_FETCH_LOGFILE
+
+ATLAS_GEOCODE_LOGFILE=$LOGS_BASEPATH/otd_geocode_atlas-lines-$DATE_NOW.log
+npm --prefix $DIR/../atlas-geocode-lines/ run process_node18 2>&1 | tee $ATLAS_GEOCODE_LOGFILE
+symlink_latest $ATLAS_GEOCODE_LOGFILE