+++ tools/gtfs-static-db-importer/CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG gtfs-static-db-importer
 
+19.Jan 2025
+- adds `trips.stop_times_count` SQLite field + index needed by the [gtfs-query](https://github.com/openTdataCH/OJP-Showcase/tree/develop/apps/gtfs-query) app
+
+09.Jan 2025
+- fix `stop_times.stop_id` GROUP_CONCAT queries which might occur in unexpected results
+
 19.Dec 2024
 - imports `original_trip_id` from trips.txt - [PR #71](https://github.com/openTdataCH/OJP-Showcase/pull/71)
     - https://opentransportdata.swiss/en/cookbook/gtfs/#tripstxt